+++ app/models/form/sales/subsections/household_characteristics.rb
@@ -33,20 +33,20 @@ class Form::Sales::Subsections::HouseholdCharacteristics < ::Form::Subsection
       Form::Sales::Pages::Person1Age.new(nil, nil, self),
       Form::Sales::Pages::Person1GenderIdentity.new(nil, nil, self),
       Form::Sales::Pages::Person1GenderIdentityJointPurchase.new(nil, nil, self),
-      Form::Sales::Pages::Person1WorkingSituation.new(nil, nil, self),
-      Form::Sales::Pages::Person1WorkingSituationJointPurchase.new(nil, nil, self),
+      Form::Sales::Pages::PersonWorkingSituation.new("person_1_working_situation", nil, self),
+      Form::Sales::Pages::PersonWorkingSituationJointPurchase.new("person_1_working_situation_joint_purchase", nil, self),
       Form::Sales::Pages::Person2Known.new(nil, nil, self),
       Form::Sales::Pages::Person2Age.new(nil, nil, self),
-      Form::Sales::Pages::Person2WorkingSituation.new(nil, nil, self),
-      Form::Sales::Pages::Person2WorkingSituationJointPurchase.new(nil, nil, self),
+      Form::Sales::Pages::PersonWorkingSituation.new("person_2_working_situation", nil, self),
+      Form::Sales::Pages::PersonWorkingSituationJointPurchase.new("person_2_working_situation_joint_purchase", nil, self),
       Form::Sales::Pages::Person3Known.new(nil, nil, self),
       Form::Sales::Pages::Person3Age.new(nil, nil, self),
-      Form::Sales::Pages::Person3WorkingSituation.new(nil, nil, self),
-      Form::Sales::Pages::Person3WorkingSituationJointPurchase.new(nil, nil, self),
+      Form::Sales::Pages::PersonWorkingSituation.new("person_3_working_situation", nil, self),
+      Form::Sales::Pages::PersonWorkingSituationJointPurchase.new("person_3_working_situation_joint_purchase", nil, self),
       Form::Sales::Pages::Person4Known.new(nil, nil, self),
       Form::Sales::Pages::Person4Age.new(nil, nil, self),
-      Form::Sales::Pages::Person4WorkingSituation.new(nil, nil, self),
-      Form::Sales::Pages::Person4WorkingSituationJointPurchase.new(nil, nil, self),
+      Form::Sales::Pages::PersonWorkingSituation.new("person_4_working_situation", nil, self),
+      Form::Sales::Pages::PersonWorkingSituationJointPurchase.new("person_4_working_situation_joint_purchase", nil, self),
     ]
   end
 end
